Fixed memory leak if a block is never added the list
block is allocated and left orphan untill it’s added to the list. If adfReadGenBlock() fails then block is never freed.

The fix is an added call to free() in that case but also something to make sure that ownership of block is always transfered to the list so that it’s not freed twice. This also simplifies the final cleanup if().