salvage: make support for salvage on volumes with dircache optional.
Salvage (ie. undeleting files) on volumes with dircache is
not fully implemented - in case of an error updating the cache
not all changes made on the volume will be reverted and
the volume might be left in inconsistent / invalid state.

That is why the support for salvage on such volumes is made
optional (and disabled by default) - and it should be like this
until full support is implemented.