regtests: add a test for checking updates of dataSize headers in OFS data blocks.
The dataSize field in OFS datablocks (only OFS volumes) were not updated
properly when the data was being appended to it. To be precise: the problem
occured when a file was written with the last data block not full, then closed,
reopened, sought to the EOF and then written with a chunk of data bigger
than the free space in the last block (so that another block is allocated).

There is already a fix for this problem (see: c0fa4799), now adding also
the regression test.

In addition, (since it is going through OFS data block headers anyway),
the test is checking also consistency of the checksums.