adfFileCreateNextBlock: (minor) bugfix (OFS).
The dataSize in newly created (empty) data block was set to blockSize
instead of 0. In practice, this was corrected immediately if any data
was put into block, as the field became updated with the proper value
so it was not causing any issues (it must have been the case, since
the new block could be requested only by adfFileWrite when needed
for data). Anyway, for consistency, it should be initialized properly.

(see also related: c0fa4799)