regtests / test_update_datasize_ofs: uncomment verifying written data.
This relates to 4ab09f7e. While data apparently was written correctly
and its verification was not a part of the issue fixed with c0fa4799,
there is a lack of the proper test verifying data after appending in
more complex cases (the tests/test_file_append.c is very basic, just
test if appending is working at all writing single bytes).
Since the data verification code is ready here (overfilling test was
the starting point), just enabling it allows to cover that, too.