adf_dir / adfEntryPrint: return pointer to the info instead of printing it.
This is more suitable for a library, client code can decide how
to use the information string.

Also, renamed to adfEntryGetInfo to better reflect the purpose.

Also - this function is widely used in tests (regtests). Since
now it requires also freeing memory, a wrapper was placed in
a common module and changes were done accordingly.