Tidy-up/refactor global lib. functions (init, cleanup, versions).
These functions were so far placed in adf_env. It was however
rather confusing in basic usage.

There are 2 new functions (adfLibInit/CleanUp) that should be
used instead of adfEnvInitDefault/adfEndCleanUp to initialize
the library and clean-up after use.

The functions for getting version info and the internal check
are moved to the new module.

All programs are updated accordingly.