adf_dev / geometry: do not fail on non-valid size devices.
Normally, every device should have size in blocks, so the total
size should be divisible by the size of the block (512).
However, there are many HDFs (hard disk files - "large
floppies", devices larger than floppies but having only one
volume on the whole device, no Rigid block etc.) having
nominally invalid sizes (non-divisible by block size,
like 5000000, 7000000 etc.).

So far, this situation was caught by adfDevIsGeometryValid_()
and opening such a device was failing and it was not possible
to perform any operations on such devices.

This modification changes the code to only issue a warning
if such "non-compliant" device is used and it can be opened
and used normally (except, of course, for the part at the very
end, after the last 512-bytes long block, which cannot be
accessed by block operations).