adfMountFlop: use value from bootblock trying to read rootblock.
So far, the rootblock's sector number set in bootblock.rootBlock
was ignored and the code was directly use the calculated value.

With this change, first, if valid, the value from bootblock is used.
If this fails, then the calculated value is used. If both fails,
the mount function returns an error.

Also, after successfully reading the rootblock, the values of
rootblock from the bootblock and the calculated one are compared.
If they differ, this means that the one in bootblock is invalid
(since the rootblock was successfully read from calculated location).