adfMountHdFile: calculate the last block from hdf's geometry.
So far, it was calculated from the position of the root block.
Since the hdf is just one big volume (like a floppy), its size
(hence - the last block) can be (better) calculated from
the device geometry.

In fact, in general case, if root block could be somewhere else
than in the middle of a volume (theoretically it can be pointed
anywhere in bootblock), the calculation of the last block relying
on root block's position would be invalid.

While the geometry of HDFs (as image files) is somewhat symbolic
(there are no physical tracks, heads or sectors), it does imply
the device's size in blocks.

There is one thing to consider - if there are cases when the hdf
is, for some reason, not used as a whole (eg. due to "bad sectors"
at the end), than the calculation based on the root block's location
would make more sense.
This, however, is an unlikely scenario, as HDFs are primarily
used as disk images (files). And not sure if there are even any
tools allowing to create such volumes (on a part of HDF).
Keeping the old code commented out, though...