examples / adfinfo: enable native driver.
This allows this utility to use the native system devices
using the interface proper for the operating system.

Normally, accessing real devices opens a way to make serious
damage to the data (globally, on the device level!), but this
program accesses devices exclusively in read-only mode (only
to read metadata information). So it is a good way to start
testing the "native" device interfaces.