Reogranize tests: move all to tests/ and make test data images common.
The tests were in 2 paths: regtests/ (regression) and tests/ (unit/functional).
Data for regr. tests was in subdirectories of regtests. Those are mostly
disk images, some of them could be possibly used in other (ie. functional)
tests.

The new layout is:
- tests/regr - regression tests
- tests/unit - unit / functional tests (using Check framework)
- tests/data - data files for tests

(In fact, there is also the 3rd path with tests: examples/tests. Since those
are related to the command-line utilities, keeping them there.)