adf_bitm / adfCountFreeBlocks: bugfix (make it work for volumes on partitions).
The prior calculation was (incorrectly) using block offsets of the device(!),
while it operated on the volume level (it should use volume's relative offsets).
It was working fine for volumes starting at the very beginning of the device
(ie. for floppies and hard files/hdfs). However, it was not correct for any
volume not starting at the beginning (volumes on a partitioned device).