adf_dev: store size in blocks, not in bytes.
Storing device size in bytes is problematic for many reasons,
major one is hitting int number limits and balancing on the edge
of overflows.

While the access to device should be always based on 512-byte blocks
and their offset, there is little reason in using byte size for
the device. In case of non-standard (or, in fact, invalid) exceptions,
like HDFs with sizes not aligned with blocks of 512 bytes, their last
part over the last 512 block is inaccessible anyway.

Using just blocks as device size simplifies a lot and makes the code
safer and more robust.

Also, added adf_limits where const values with size limitations are
stored. They may depend on platform (as described in comments there).