adf_dev: refactor opening and mounting, make RDB part of the state.
Opening the device and setting device states (type, class, geometry) is
somewhat complicated. This modification is an attempt to improve it a bit.

There are 2 major things which setting should (hopefully) be more clear:
- checking RDB's existence and, if found, setting geometry from RDB
- setting device class (ie. ADF_DEV_CLASS_HARDDISK vs. _HARDFILE) depending
  on presence of RDB (so, in consequence, also partition table etc.)

The state of RDB, as well as RDB block itself (optionally, can be read when
needed), is now part of the dev. state (see adf_dev.h).
(It can be useful also in the client code.)