tests / examples / common: ensure test script does not end with a disaster.
The command used in the script for cleanup (in common.sh):

trap cleanup 0 1 2
cleanup() {
    rm -rf $tmpdir
}

can potentially end in an utter distaster. $tmpdir can be changed
in any of the scripts the scripts sourcing the discussed common.sh,
then, modified, is used by cleanup(). If, for any reason, $tmpdir
becomes empty (not that bad, yet...) or "/" (unlikely, but POSSIBLE)
- the result requires no comments...

Extermely dangerous commands like "rm -rf" must be used in scripts
with absolute caution. "rm -rf" should only be used on contents of
_hardcoded_ subdirectories,