tests / examples(2): add first test of adfinfo with a hard disk dump.
Also, this test is done within newly developed framework for
command-line tests. The idea is to make something more strictly
structured but what ultimately should be easier to maintain
(ie. able to test but also to generate expected tests results,
which manual updating is a chore...), at the same time also working
with both CMake and autotools, as well as easy to execute manually
(selectively).
So far this is a somewhat working prototype. If it is gonna work as
expected - that is still to see...